src/pyproject.toml
[tool.poetry]
name = "datayoga"
version = "0.0.0"
description = "DataYoga for Python"
authors = ["DataYoga <admin@datayoga.io>"]
license = "Apache-2.0"
readme = "README.md"
include = [
]

classifiers = [
        'Development Status :: 4 - Beta',
        'Intended Audience :: Developers',
        'License :: OSI Approved :: Apache-2.0',
        'Programming Language :: Python :: 3.7',
        'Topic :: Database'
]

[tool.poetry.dependencies]
jmespath = "^1.0.0"
jsonschema = "^4.4.0"
python = "^3.7"
PyYAML = "^6.0"

mock = { version = "^4.0.3", optional = true }
pytest = { version = "^7.1.2", optional = true }
pytest-describe = { version = "^2.0.1", optional = true }
pytest-mock = { version = "^3.7.0", optional = true }
pytest-timeout = { version = "^2.1.0", optional = true }
requests-mock = { version = "^1.9.3", optional = true }

[tool.poetry.extras]
test = ["mock", "pytest", "pytest-describe", "pytest-mock", "pytest-timeout", "requests-mock"]


[tool.poetry.urls]
url = "https://datayoga.io"
[tool.poetry.scripts]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"